Bug 699628: Add fz_var to variable set in fz_try and used in fz_always.
When pdf_xobject_transparency() indicates transparency,
pdf_run_xobject() will push a mask with 2 components (gray +
alpha). Because fz_var was missing for the transparency variable
its value would, if fz_throw() was called, not be preserved in
fz_always{}. This means that fz_always{} would never pop the mask
from the stack. At a later stage some code would call
fz_end_mask() and then trigger an assert() where the number of
components was restricted to 1.

Marking the variable with fz_var means that the mask will now be
popped as expected. Thanks to oss-fuzz for reporting.