Bug 699631: Handle unsupported triangle mesh edge flags.
There were two issues with the code parsing the triangle mesh's
edge flags:

  * meshes were not require to start with an independent triangle
  * out of range edge flags caused vertices to be ignored

A mesh where the edge flag of the first vertice is out of range,
and the edge flag of the second vertex indicates continuation
of a prior triangle would result in trying to create a triangle
where the third coordinate would be uninitialized.

This commit requires the edge flag of the first vertex to indicate
a new independent triangle and if out of range edge flags are
encountered they are treated as if they indicate a new triangle.

Thanks to oss-fuzz for reporting.