Fix 700053: Add 'decrypt' option to pdf_save_document.
This option is exposed as the -D flag to mutool clean.

Allow saving a document without encryption.

The next step is to allow saving a document with new encryption, but this
may require a fair amount of rejigging since the PDF write code to keep
separate doc->crypt objects for decryption and encryption.