Add a form-editing flag to the widget state.
This is used to allow text widgets to be updated character by character
without any sideeffects on other widgets or running of javascript events.
This can be set while performing in-place editing. At the end of editing
the field, the state should be reset and a final seting made to the text
to perform the auxiliary actions.