java: Modifications for supporting form filling.
Allow PDFDocument to enable javascript support. When this is done a
PDFDocument.JsEventListener should be supplied, this contains callbacks to
be called upon events, e.g. alerts.

PDFWidgets was added as a new class providing the needs for form filling,
e.g. getting/setting the widget value, content type, maximum length, etc.

PDFPage adds a member that returns an array of all widgets on the page,
called getWidgets(). PDFPage also adds selectWidgetAtNative(x, y), which
clicks on the widget at the given x, y coordinate and also returns the
widget that was clicked.