Fix bug in toggling of checkboxes and radio buttons.
We weren't properly handling the unchecking of ganged buttons (ones which
share the same on-state), and in fact, we were handling the whole task in
an unnecessarily complicated way. The way radio buttons and checkboxes are
handled are almost identical, the difference in behaviour coming from the
grouping and the naming of on states.

There is one case we don't currently handle correctly, which is a case
where the behaviour has changed between different Acrobat versions. It is
where a group is marked as being radio buttons but all the on states are
the same. We treat these as ganged. We should do so only if the
radiosinunison flag is set and otherwise special case them to be one on
at a time. I haven't attempted to address this case because I cannot find
an example file to test under Acrobat to see how it should be handled. We
didn't address that case correctly before this commit, in any case. I wrote
a program to detect this case, but didn't find an example file when
scanning my local collection of forms.