Fix Images that use matte with subarea and/or l2factor.
Running the test file from bug 700571 through MuPDF results in a
warning about a mismatch between the mask and the image sizes for
an image with a "Matte" mask.

The code in MuPDF for dealing with such mattes relies on both
the image and the mask being the same size. The first thing that
defeats this, is that any use of subsampling can result in
different sizes. We therefore disable native subsampling (as this
happens before the matte application), and rely on post decode
subsampling.

Secondly, the matte logic needs to be updated so we can cope if
we are only decoding a subrectangle of the image. We request a
matching (or larger) subarea of the mask, and read the correct
area of that to mask with.