js: Add util.scand and clean up util.printd.
util.printd does not print a timezone (contrary to spec), so
remove the timezone from the default formats.

util.printd also checks the date argument for truthiness,
and if it is false, substitutes the current time.
If the date argument is not a date object, it uses it as the argument to the
Date constructor.

Implement util.scand by matching the format string accurately.
This is what the reference states, but Adobe seems to be using
an algorithm closer to AFParseDateEx, which basically just ignores
the format string for all but getting the y/m/d order.

Remove the MuPDF object holding various constants, since they
are only used in a single function at a time, we can clean up
and make them local.