Avoid dropping object twice upon exception during repair.
In pdf_repair_xref() /Root is obtained and then add_root()
is called. If add_root()s own call to fz_resize_array() fails
due to running out of memory the /Root obj variable is dropped
in the fz_catch() at the end of pdf_repair_xref(). This is a
problem because when /Root is obtained and stored in the obj
variable it is only a borrowed reference, so not a reference
that needs to be freed. On the other hand the same obj variable
is elsewhere in pdf_repair_xref() used to store newly created
objects whose references do need to be dropped.

To resolve this the two uses of the obj variable is separated
so that the reference stored in the obj variable always can be
dropped, while the dictobj variable is never dropped since it
does not store object references that need to be dropped.