Bug 697655: Mark all objects seen during page separation scan.
Keep a list of objects to unmark at the end of each pass.
This prevents us scanning the same resource dictionary
multiple times, if we've already seen it.

Ignore broken colorspaces in all places, not just the ColorSpace
resource dictionary.

This also incidentally fixes bug 700720:

Pages and XObjects may share dictionaries, without that necessarily
leading to actual recursion in their use. Allow for this by stopping
when encountering a resource dictionary we've already scanned instead of
throwing an exception.