Fix strings with null chars being formatted inefficiently.
Recent changes, analyse strings to determine whether they are more
efficiently stored as characters or hex, but the requirement for nulls
were being miscalculated due to strchr including the terminating null
in its search. Bringing the test that was supposed to capture null earlier
in the sequence of tests, fixes the problem.