Purge unmaintained progressive loading code.
The progressive loading and trylater mechanism adds a substantial
amount of complexity and special error handling to allow the
device interface to proceed in the face of exceptions.

The progressive loading only has a purpose for linearized files,
and has issues where it can get stuck in a state and never
progress. Since none of these issues have been reported by
users or customers, we can assume that nobody uses this
feature.

This code is notoriously fragile and difficult to get right,
and is a constant source of oss-fuzz bugs and assert errors.

Let's reduce the technical debt by removing this feature.
This commit removes the progressive loading and trylater
exceptions. Following commits will clean up and remove
the error handling that is no longer necessary.