Recast unhandled syntax errors as generic errors in pdf_run_xobject.
Since we handle syntax errors in pdf_process_contents, none should escape
from pdf_run_xobbject; but in some cases auxiliary parsing of dictionaries
could cause some of these to be thrown. These are generally fatal, and are
not recoverable, so should not be passed up to the caller with the
FZ_SYNTAX_ERROR type.