Bug 700913: Limit number of sub function outputs for stitching functions.
Previously when evaluating stiching functions it was assumed that
the number of outputs for each sub function was the number of
outputs of the respective sub functions. This is not correct, as
the number of outputs for each sub function must be the same and
limited to the half the number of entries in the /Range array of
the stitching function.

That is, a stitching function determines the number of expected
outputs of each sub function. Should the sub function supply too
few or too many outputs, these should be assumed to be zero or
ignored respectively. After this commit this is what happens.