Retain PDF processor default colorspaces upon early exception.
If an exception occurred while running an xobject prior to
save_default_cs was set then the default colorspaces for the PDF
processor would always be reset to NULL, If the exception
occurred after save_default_cs was set then the default
colorspaces were retained as intended.

Now the default colorspaces are retained regardless of where the
exception occurs.