java: Allow 'null' PDFObjects.
The paranoia test for whether an object has been destroyed also triggers
on the PDFObject.Null instance, which has a NULL native pointer. The
destroy method is not really a necessary part of the API, and dispensing
with it will solve this issue without needing to add another field to
track liveness of the pointer.