Pass the selection/highlight character bbox to the text filter.
Pass a bounding box that matches the same area that would be shown
by the stext-device when selecting or highlighting text. This covers
the glyph origin and advance width and line height, as opposed to the
inked area of the glyph which is only useful for rendering purposes.

Also pass the TRM and CTM so that the callee can do whatever transforms
it deems necessary. The TRM holds the font size and position in text
space, and if concatened with the TRM you can get the PDF page space
coordinates and size.

Further concatenated with the page transform, you can also get device
coordinates.