Bug 700692: Pass scissor rectangle into shade handling.
This is required to ensure that the rectangle we create for linear
fills adequately covers the required area.

In theory, we could get called with no scissor rectangle, if
someone were to invoke a shading at the top level with no clip
path, but a) this falls back to our previous maths, so won't
crash, and b) shouldn't ever happen in practise.