Bug 700692: Follow on to previous fix for 'extend' regions.
The customer points out, correctly, that while the shade now
appears, the color is not right. This is due to the 'extend'
regions of the shade overflowing, and becoming unrepresentable
in ints. Consequently the wrong 'extend' region is plotted
giving the wrong color.

The fix is to change the extent of the 'extend' regions to
be just enough to cover the scissor region, rather than being
'HUGE'.