Use correct alpha component index when converting separation pixmap.
The order of components is: [processcolors spotcolors alpha]
In the code sd is set to point to the first spot color in the source.
Previously the code attempted to obtain source alpha by doing sd[sc].
This would have been true if sd pointed to the first proces color
and there were not spotcolors in the source. To get to the alpha
component all spot colors must be skipped over, and since sd points
to the first spotcolor then you need to do sd[ss] to get the source
alpha.