Fix Matte rendering issues with subareas.
The following test:

   tests_private/pdf/PDFIA1.7_SUBSET/CATX4135.pdf.ppmraw.300.0

is showing problems with a missing yellow band. This renders fine
at < 100 dpi.

Investigation shows that this is to do with the matte calculations
being badly offset if we get a full image cached when we expect
a subarea to be cached.

Fixed here, in what should be a safe, but aesthetically unpleasing
way.