Generate smaller builtin fonts with better subset.
Update and add URW Type1 source fonts.

Add CharisSIL source fonts.

The script makesubset.py loads and prints a list of glyph names suitable
for using with the 'tx' tool's -g option to do the subsetting.

We create several lists of possible characters, based on various subsets
of unicode:

	* MES-2 (Multilinguan European Character Set)
	* SECS (Simple European Character Set)
	* WGL4 (Windows Glyph List 4 for the Pan-European character set)

The subset used to generate the fonts in this commit is the MES-2 list.
If you want even smaller font files, use the WGL4 or SECS lists instead,
by editing and rerunning the tocff.sh scripts.

For CharisSIL regular we also add small-caps variant glyphs for all the
small letters and ligatures.

https://www.cl.cam.ac.uk/~mgk25/ucs/mes-2-rationale.html
https://www.cl.cam.ac.uk/~mgk25/ucs/secs.html
https://docs.microsoft.com/en-us/typography/opentype/spec/wgl4