Bug 701118: Handle appending large display list nodes.
The size of the begin layer node depends on the size of the layer
name. That name may be a string from the page's property resources,
and is only bounded by memory when parsed by lex_string(). So the
append_list_node() logic cannot simply double the size of the
display list and hope that the node fits, since the node may be
of arbitrary size.

Now append_list_node() would repeatedly double the size of the
display list until the node fits, or malloc() runs out of memory.