Bug 701118: Limit size of begin layer nodes in display list.
The size of the begin layer node depends on the size of the layer
name. That name may be a string from the page's property resources,
and is only bounded by memory when parsed by lex_string(). The
layer name may cause a display node to be larger than the maximum
size allowed. This condition is now checked for.