Bug 701170: Restrict number of colorants for pixmaps.
Previously the number of colorants was unrestricted. Consider the case
where a CYMK pixmap without alpha but with 30 separations was used as a
destination. This means that the total number of colorants exceeds
FZ_MAX_COLOR. If one attempted to draw graphcis the set color would need to
be converted to the destination colorspace and separations, overwriting the
bounds of any involved color arrays.

Now the number of colorants and alpha is restricted to FZ_MAX_COLORS for
any created pixmap.