Improve redaction capabilities.
When redacting text from a PDF file, we want to ensure that we
redact the same text in the StructTree parts.

Add 'structparents' param to the text filter creation; this should
be the number given in the content stream's "StructParents" entry
(or -1 if there isn't one).

Within the PDF filter processor itself, we move to keeping track
of current tags as well as pending ones. As we encounter each tag,
we look for the /MCID object, if there is one, and we store a
reference to this in the tag_record. From this, we look for /Alt
and /ActualText tags, and, if present, extract the strings from
them as UTF8.

Whilst filtering, whenever we encounter a char, we consider the
/MCID's that are in force; if we have Alt or ActualText strings
we check whether the next character agrees with the character
pulled from the file or not. If the strings agree, and we have
to remove a character, we remove it from the Alt/ActualText strings
too. If the strings disagree (and characters are removed), then
we err on the side of caution and remove the entire string.

When we come to end the marked content (with an EMC), we update
the edited strings in the file.