Add a function to check for incremental changes since a signature.
There are two ways in which a document may change after the signing of
a signature field: either the body of the document at the time of signing
may be altered, or some edits may be saved by incremental update. We
already have a way to detect the first case via checking the digest
against the signature. This new function allows us to detect the second
case.