Validate signature dictionary byte ranges.
The discontinuous byte ranges are sorted and assumed to cover the
entire PDF file. Any holes are assumes to be certificate
signatures and must therefore contain hex byte strings according
to specification. If any byte range extends outside of the file
or any hole does not entirely consist of hex digits, then the
signature can not be considered to be valid.