Use signature dicts directly, not widgets, in signature functions.
Fields in /Root/AcroForm/Fields are not required to refer to the
page they are used on. This means that they can not be treated as
widgets, who have to refer to the page they are used on.

Iterating over /Root/AcroForm/Fields to find signature fields is
faster than iterating over all pages to find signature widgets.
Functions operating on signature dictionaries directly, can be
used both with signature widgets and signature fields, allowing
for this quicker signature location method.