Solve Type3 font with zero sized glyphs being very slow.
This problem was seen with customer file 177b.pdf. The type 3
font contained some empty glyphs. These had an empty bbox.

The current code distrusts empty bboxes, and fall back to the
font bbox, resulting in huge (9000x9000) glyphs which take an
age to render and don't cache.

Here we fix the code so that if we have an empty bbox, we find
the accurate bbox from the glyph. If that happens to be empty
we store it as a tiny rectangle to avoid checking every time.