Remove unneccessary output context.
The ability to set custom fz_output streams as fz_stdout/stderr was intended as
a way to catch error and warning messages. Due to possible unwanted recursion
(printing the error message throws an error that prints an error message that
throws an error, etc) we never got around to using the customizability in this
way.

Since users can log error messages via the error/warning context call backs,
this unneeded functionality can safely be removed.

Since we now disallow changing the stdout output stream, we can also simplify
the handling of the fz_stdout stream.