Fix for unencrypted signature dictionary Contents entry.
Add a load_unencrypted_object function to load an indirect object without
applying decryption. We use this to load the signature Contents string
without encryption, since per spec (pdfref20 page 82, 7.6.2) this is
one of the few strings that should not be encrypted.

The Encryption dictionary and ID entries are loaded before we create the
decryption structure so these cases are already handled.