Bug 701311: Avoid accessing off end of array.
When scanning a fuzzed file for linearisation, we were encountering
a page object (4118) that was larger than the number of objects
in the file (~1200). We put that number in the page object list,
and then later tried to adjust that number for renumbering by
looking it up in the renumber map. Because the renumber map wasn't
long enough, we accessed off the end.

Either this could cause crashes, or the fact that that then brings
in bogus data can later cause crashes.

The fix implemented here is to ensure that all the lists are
large enough to encompass the page objects we put in.