Use round caps/joins when generating ink annotation appearances.
This makes ink annotations look less jagged and allows for
degenerate annotations with a single point to be visible.

Make sure that generated appearance streams for single point
ink annotation contain a moveto operator and a lineto operator.
This means that when the path is stroked it will look like a dot.