Optimise memory use in fz_xml.
Store one less pointer; tail is only used during creation, at a
time when 'next' isn't being used. Remove it and use next instead.

Move names to the end of the structures, so the structures can be
truncated to the actual size required. This has the extra
benefit of no longer restricting us to 40 chars for a tag or
attribute name.