Set signature content end marker to /Type /Sig.
In commit a56db95ef we added the optional '/Type /Sig' entry to the signature dictionary,
but the code in pdf-write.c:complete_signatures() was still using the '/Filter' entry as
the marker for the end of the digest content. This resulted in the calculations for the
byte range being out by 8 as we were including the recently added /Type /Sig entry, and
resulted in verify errors when we parsed the digest in pdf-form.c:validate_certificate_data().

This commit modifies pdf-form.c:pdf_signature_set_value() to add the /Type entry after the
other entries so that the mandatory /Filter entry is still directly after the contents.