java: Implement helper functions for ByteBuffers in jni.
We use direct buffers where possible when digitally signing a document.
Unfortunately we can't guarantee that this functionality will be available on
all devices or implementations, so we also need to cope with bytearray-backed
implementations which will typically require some copying.

Add some utility functions to manage access to a plain byte buffer representation
of a ByteBuffer object.