Fix store behaviour when oversized.
The store can become oversized while objects that are in it are
in use. The idea is that once they cease to be in used, the
store should clear out enough to keep it below the specified
maximum level. The current code fails to do this.

When the store is oversized, if we drop a storable item and it
ends up so that the only reference to it is from the store, we
should bin it entirely so the store can shrink.

Update fz_drop_storable to perform in this way.