XML parser: Tweak for improved HTML parsing
The XML parser copes badly with HTML, for several reasons.

Firstly, many HTML tags close themselves, without requiring
a "/>". (e.g. <meta foo=blah>). The XML parser would leave
such tags open.

Secondly, many HTML tags should close themselves when a
second sucessive tag of the same type is found (e.g.
"<p>foo<p>bar"). Again, the XML parser would leave such
tags open.

For situations such as tables, this is even worse; for example
a TR tag should automatically close any existing TD tags.

Thirdly, HTML is case insensitive, so tags need not exactly
match.

Finally, the XML parser itself will get confused if the XML
is less than perfect. e.g. <a><b></a> will see the </a> close
the <b> tag, but not the <a> tag.

We therefore introduce some improvements to the parser to cope
with such issues.

The biggest change is a "for_html" flag to the xml parser; this
triggers various changes to the way the parser works.

Firstly, it lowercases all HTML tags as they are read.

Secondly, it makes use of a table of all defined HTML tags with
various flags so that any tags that autoclose themselves (or other
tags) are handled correctly. In addition, using the same table,
we can 'auto open' tags (for instance if we meet a <tr> without
a <table>, we can automatically open the <table> first).

This makes us much more resilient to HTML (as opposed to XHTML)
files.

In both XML and HTML parsing modes, we now check the tag we are
closing, so <a><b></a> will close both the <b> and <a> tags when
the </a> is parsed. Unmatched closing tags are ignored, so
<a><b></a></b> will still be parsed correctly.

All of these changes should leave operation on well formed XML
files unchanged.