bmp: Improve run-length decoding to better handle bogus images.
The RLE decoders now keep track of an X,Y coordinate.

The Y coordinate is only updated when End of line or Delta codes
are encountered. The X coordinate is updated when processing any
type of code. In addition the X,Y coordinate are allowed to
exceed the image dimensions.

Only when encountering the codes Absolute or Encoded that unpack
pixel samples is the X,Y coordinate mapped to a particular pixel
in the decoded image. X coordinates that exceed the image width
are wrapped to the next pixel row, even if this is not reflected
in the Y coordinate. This is to support images that completely
lack End of line codes. If the Y coordinate exceeds the image
height no pixel samples are updated in the decoded image.

The decoders now also attempt to decode as much as possible, even
given only partial data necessary to process a code. E.g. when
encountering an Encoded code that wants to unpack 10 pixels, but
data remaining may only decode 5 pixels, then the 5 pixels are
decoded.

This fixes Coverity CID 125729.