bmp: Remove warning when RLE skips decoding last pixels in image.
RLE compression allow skipping over output pixels when decoding.
MuPDF leaves these pixels as palette index 0 for RLE4 and RLE8,
or black for RLE24. This applies also to pixels at the very end
of the image, which may be skipped over not by using Delta codes,
but rather through an End of bitmap code. MuPDF should not warn
in this case.