Add functions to validate changes within a PDF file.
1) Add pdf_locked_fields API.

pdf_find_locked_fields() finds the fields that are locked in a
given version as an opaque structure.

pdf_drop_locked_fields() drops that structure.

pdf_is_field_locked() checks whether a given named field is locked.

2) Add pdf_validate_changes API.

This works on the principle of making a list of ALL the changes
between 2 versions of the xref, and then atetmpting to justify
those changes. Any unjustified changes will cause the validation
to fail.

To do this we need to know how many 'updates' there have been to
a document - we introduce a function pdf_count_incremental_updates
for this. The only wrinkle here is to do with linearized files.

Linearized files are initially written as a short file, plus an
update. The short file is not complete, as it refers to objects
in the 'update'. We therefore have to special case this; we spot
linearized files, and count those as being one 'atomic' write
of the file, so we don't count that as one of our incremental
updates.

This requires us to make changes to the PDF loading code to spot
linearized objects in all cases (whether we are loading them
linearly or not).

To find the trailer of a document, we call pdf_trailer(). Until
now this has not been honouring the xref_base value (so we always
got the most recent trailer, not the one for the specified
update). Fix that here.

3) Add pdf_validate_change_history() function.

This calls the previously introduced API to check that all the
changes made to the document are consistent.