Bug 701977: Workaround for bug 57519 in FreeType.
FT_Get_Advance has a bug with certain Type1 fonts, because the fast
metrics parsing function does not handle 'div' operators.

Disable the fast metrics for Type1 by forcing the use of the old
Type1 engine for our builds. This will not work for builds using
the system library, and this workaround should be removed as soon
as we update to a FreeType a release with a fix for this bug.