Bug 701754: Check 'uniXXXX' glyph name length to avoid false matches.
A prefix of 'uni' is not sufficient, as evidenced by the file attached
in the bug report. Reject glyph names that start with uni but are not
7 characters long, to allow us to fall back to other last resort
approaches to encoding/decoding text.

In the example file we'll encode these glyphs as U+FFFD replacement
character, since we cannot find a good encoding for them.