Update pdf_walk_tree to make it more general.
Frequently, when we are walking trees, we want to refer to
values that might be inherited from parents. Unfortunately,
we can't call pdf_get_dict_inheritable, because we will have
been 'marking' the tree on the way down, and this will
conflict with pdf_get_dict_inheritable marking the tree on
the way up.

We therefore add a mechanism to pdf_walk_tree, where we give
it a list of names and it can collect the values of those
as we move down the tree.

These are then passed into our callback function so we can
use them when we get to the nodes.

We also update the code so that we have 2 callback functions,
rather than one. The first one is called on arrival at the node
(prior to heading down to any kids). The second one is called
just before leaving a node.

This should allow us to do stuff like accumulating names as we
pass down the tree.