Move pdf incremental saving file copy logic into library.
If we try to write a pdf file with 'incremental' set, and we're
not writing to the same file (which is risky, and illadvised), we
need to copy the file data across from the old file first.

Previously this had to be done in the app. Move that logic into
the library.

We rely on being able to get an fz_stream from the fz_output to
be used; if we can't get one, then the incremental write will
fail.

This requires a slight change to fz_new_output_with_path when
called with append=1. Previously we tried to open the file with
"rb+", which fails if the file does not exist. In this case we
now retry with "wb+". This means we can safely "append" to a
non-existent file.