mupdf-gl: Force a save of the document after signing.
When we're about to sign the doc, trigger the save dialog,
and actually do the signing just before the save.

This avoids us ending up with a signed document in memory
that doesn't get saved (which will then cause problems if
any other edits happen to it).

Whenever we save a document, reload from it; this is vital
to allow stuff (like signature validation, and future saves)
to work.