Tweak pdf_validate_changes/version counting.
1) pdf_validate_changes:

Avoid "accepting all changes" at any point, as this risks being
too permissive for maliciously constructed files. Instead, introduce
a function, filter_changes_accepted, that will recursively walk an
object accepting changes in non-indirect children, and only stepping
down into indirect children if they are 'expected' by the filter.

Secondly, if no fields are locked, and no document permissions are
set, then allow any changes. This copes with the case seen in
af171.pdf, where a pure-XFA file loaded into Acrobat Reader, signed
and saved back gives a document where the page content streams have
changed as part of the signing. Without this change this was held
to be an illegal change.

Also, avoid leaking a pdf_locked_fields record.

2) Simplify PDF version counting functions:

Within MuPDF we number the 'incremental updates' we've had in
a file from 0 (the latest) to n (the oldest).

Working from 0, we have:
   * any unsaved changes
   * the last incremental update from the file
   * the penultimate incremental update from the file
   * ...
   * the first incremental update from the file
   * the 'base' file version.

If a file was originally written linearised, then the 'base' + 'first
incremental update' should really be considered as a single atomic
version of the file.

All this becomes hard to reason about, so we remove
pdf_count_incremental_updates in favour of 2 new functions,
pdf_count_unsaved_versions() and pdf_count_versions() that
return the number of unsaved and saved versions respectively.