Fix possible infinite loop in pdf_encrypt_data.
The variable n was recently changed from type int to size_t. The first
loop in the function tollerated this, but the second may attempt to take
n negative. Part of the solution is to change the decrementing of n to
avoid going negative, but that change alone would have led to the
following if block always being executed. That block was supposed to
execute only for buffers with size a multiple of 16. To maintain that
condition, we have made len available outside the loop and test for the
loop ending with len being 16. (Thanks to Robin for the details of this
solution).