Bug 702158: Handle Lab when converting separation pixmaps.
Previously when a separation pixmap was converted to its base
colorspace there was a bug if the base colorspace was Lab. At this
point the converted components were assumed to be in the range 0..1.
This assumption is correct for all colorspaces except Lab where the
components normally are in the range 0..100, -128..127 and -128..127.
Lab colorspaces are now handled separately taking care to convert
component values accurately.