Fix fields not locked on signing.
The locking of fields is based on the matching of field names. It used to
work, but we later discovered that field names were being derived
wrongly and in response we fixed only one of the two places where they
were derived and thus the matching stopped working. This commit corrects
the second place where field names are derived.