Bug 701607, 701610: Tweak text redaction filter.
Using the character origin turns out to be too fragile.

In one case, we are getting false positives since the next character
abuts the quad and rounding errors cause the test to succeed.

In another case, we are getting false negatives since the text marked
consists of two overlapping lines: one line of text and one line of
drop shadow text. The slight difference in overlap causes one of the
halves to be missed when redacting.

Use the character bounding box instead, and if it intersects the redaction
area we elide the character. Shrink the character box in order to allow
for a bit of fuzz and rounding errors.