PDF forms; don't (re)generate signature appearance streams
Previously, if a signature form field had no appearance stream, we'd
generate it one. Signature fields that 'certify' a form can have
a null bounding box, and no appearance stream.

The act of attempting to generate one is enough to move the object
into the latest incremental update within the file, which invalidates
the signature.

The fix, implemented here, is simply to check whether a field is
a signature before attempting to generate. If the field is a
signature we refuse to regenerate the appearance stream.