Parse XFA and use it for locking fields in PDF forms.
It seems that lots of AcroForms are generated from XFA forms
as part of the authoring process. Sadly, this generation
process does not extend to listing the fields that should be
locked by signing a given signature.

This means that when we sign such a document (such as dd2875.pdf)
the signature does not lock the fields that you'd expect, meaning
that they can still be edited when it is reloaded.

We add a workaround for this here; upon signing, we parse the XFA
for a document and cache it in the pdf_document. We then scan
this for details of which fields should be locked by a given
signature, and ensure that those are locked.

This process *shouldn't* ever go wrong and cause us to lock more
fields than we should, but it's possible we may need to tweak it
a bit to cope with variations in the layout of files that might
mean we don't lock all the fields we should. We can evolve this
over time.

We add a pdf_invalidate_xfa() function for people to call in the
(vanishingly unlikely) event that they modify the XFA of a
document.